Phase 3a: Split ControlContextProvider into focused hooks
control-context.tsx was 978 lines and 43 hooks doing seven concerns at
once: apiKey, BYOK, secret vault, model fetch, app fetch, per-thread
selection, custom pub/sub. Adding anything meant scrolling 800 lines to
find the right effect.

Extracts the body into four focused hooks under packages/react/src/control/:

  api-key.ts            useApiKey()
    apiKey state + localStorage round-trip + setter.

  byok.ts               useByok()
    BYOK provider keys + the generic secret vault API (ingestSecrets,
    clearSecrets, deleteSecret, listSecrets). They share a file because
    BYOK actions internally call the vault, so splitting them would
    just create call-backs across the seam.

  auth-endpoints.ts     useAuthEndpoints()
    /api/session/apps and /api/session/models fetches. Effect deps stay
    quiet across thread switches (the May 2026 fix preserved verbatim;
    extensively documented in the file header).

  per-thread-control.ts usePerThreadControl()
    getCurrentThreadControl, getCurrentThreadApp, onModelSelect,
    onAppSelect, syncCurrentThreadControl, markControlSynced, and the
    auto-model-sync effect. Reads other slices via refs.

ControlContextProvider now reads as the wiring diagram it always wanted
to be: refs into the central plumbing, four useFooHook calls, the
aggregate state, and one assembly statement. Down from 978 lines to 250.

Deleted in this pass:

- onControlStateChange and the `callbacks` Set ref. No external
  consumers (verified by grep across packages/, apps/). It was a parallel
  pub/sub layer beside the React context that produced the same data —
  one access pattern too many.

- The legacy `setState({app, apiKey})` shim. Deprecated since the May
  2026 namespace -> app rename; no remaining callers in the repo.

useControl() is preserved with the same shape so consumers keep working.
The new focused hooks are also exported (useApiKey, useByok,
useAuthEndpoints, usePerThreadControl) for code that only needs one
slice. Consumer migration follows in Phase 3b.

136/136 tests pass (12 control-context + 124 runtime/session).
